moving stuff around unnecessarily, user input
